+++ pkg/handler_querydata.go
@@ -3,10 +3,13 @@ package main
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"time"
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
+	"github.com/tidwall/gjson"
+	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/framer"
 	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/infinity"
 )
 
@@ -63,9 +66,43 @@ func QueryData(ctx context.Context, backendQuery backend.DataQuery, infClient in
 		Error:                  "",
 	}
 	if query.Source == "url" {
-		response, statusCode, duration, err := infClient.GetResults(query, requestHeaders)
+		urlResponseObject, statusCode, duration, err := infClient.GetResults(query, requestHeaders)
+		if query.Type == infinity.QueryTypeJSONBackend {
+			if query.RootSelector != "" {
+				responseString, err := json.Marshal(urlResponseObject)
+				if err != nil {
+					backend.Logger.Error("error json parsing root data", "error", err.Error())
+					response.Error = fmt.Errorf("error parsing json root data")
+					return response
+				}
+				if !gjson.Valid(string(responseString)) {
+					backend.Logger.Error("error json parsing root data")
+					response.Error = fmt.Errorf("error parsing json root data")
+					return response
+				}
+				r := gjson.Get(string(responseString), query.RootSelector)
+				var out interface{}
+				err = json.Unmarshal([]byte(r.String()), &out)
+				if err != nil {
+					backend.Logger.Error("error json parsing root data", "error", err.Error())
+					response.Error = fmt.Errorf("error parsing json root data")
+					return response
+				}
+				urlResponseObject = out
+			}
+			newFrame, err := framer.ToDataFrame(query.RefID, urlResponseObject, framer.FramerOptions{}, "")
+			if err != nil {
+				backend.Logger.Error("error getting response for query", "error", err.Error())
+				customMeta.Error = err.Error()
+			}
+			if newFrame != nil {
+				for _, ff := range newFrame.Fields {
+					frame.Fields = append(frame.Fields, ff)
+				}
+			}
+		}
 		frame.Meta.ExecutedQueryString = infClient.GetExecutedURL(query)
-		customMeta.Data = response
+		customMeta.Data = urlResponseObject
 		customMeta.ResponseCodeFromServer = statusCode
 		customMeta.Duration = duration
 		if err != nil {
@@ -75,6 +112,34 @@ func QueryData(ctx context.Context, backendQuery backend.DataQuery, infClient in
 	}
 	if query.Source == "inline" {
 		customMeta.Data = query.Data
+		if query.Type == infinity.QueryTypeJSONBackend {
+			data := query.Data
+			if query.RootSelector != "" {
+				if !gjson.Valid(data) {
+					backend.Logger.Error("error json parsing root data")
+					response.Error = fmt.Errorf("error parsing json root data")
+					return response
+				}
+				r := gjson.Get(data, query.RootSelector)
+				data = r.String()
+			}
+			var out interface{}
+			err := json.Unmarshal([]byte(data), &out)
+			if err != nil {
+				backend.Logger.Error("error json parsing", "error", err.Error())
+				customMeta.Error = fmt.Errorf("error parsing json %s", err.Error()).Error()
+			}
+			newFrame, err := framer.ToDataFrame(query.RefID, out, framer.FramerOptions{}, "")
+			if err != nil {
+				backend.Logger.Error("error building frame", "error", err.Error())
+				customMeta.Error = err.Error()
+			}
+			if newFrame != nil {
+				for _, ff := range newFrame.Fields {
+					frame.Fields = append(frame.Fields, ff)
+				}
+			}
+		}
 	}
 	frame.Meta.Custom = customMeta
 	response.Frames = append(response.Frames, frame)