+++ src/editors/query/query.json-backend.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { InfinityQuery, EditorMode } from '../../types';
+
+export const JSONBackendEditor = (props: { mode: EditorMode; query: InfinityQuery; onChange: (value: InfinityQuery) => void; onRunQuery: () => void }) => {
+  return <></>;
+};